requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

httpx>=0.26.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dateutil>=2.8.0

apscheduler>=3.10.0
loguru>=0.7.0
fake-useragent>=1.4.0

pytest>=7.4.0
pytest-asyncio>=0.23.0
